requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
builtwith==1.3.4
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.5
django-colorfield==0.4.2
django-storages==1.11.1
django-mathfilters==1.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
msgpack==1.0.2
Pillow==8.3.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.6
whitenoise==5.2.0
zope.interface==5.4.0
